pyproject.toml
[project]
name = "warwick-wrb-checker"
version = "0.1.0"
dependencies = [
    "requests",
    "python-dotenv"
]
requires-python = ">=3.10"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov"
]
